Add v1 of the SDK (#2)
This adds the v1 of the SDK for interacting with the Multi-Collateral Mento contracts. The current version is built under two assumptions:

Each exchange has exactly 2 assets
There's only one exchange for a given pair of assets

While I know both of these assumptions won't be true in the future I thought it didn't make a lot of sense to overcomplicate things at the moment when we won't have them at launch. Nonetheless the code should be structured in a way which should make extending to more assets and exchanges relatively easy.